Splits crates into "dep"s to remove cfg feature "program".
