+++ programs/sbf/rust/invoke/Cargo.toml
@@ -8,15 +8,12 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[features]
-default = ["program"]
-program = []
-
 [dependencies]
 rustversion = { workspace = true }
 solana-program = { workspace = true }
-solana-sbf-rust-invoked = { workspace = true }
-solana-sbf-rust-realloc = { workspace = true }
+solana-sbf-rust-invoke-dep = { workspace = true }
+solana-sbf-rust-invoked-dep = { workspace = true }
+solana-sbf-rust-realloc-dep = { workspace = true }
 
 [lib]
-crate-type = ["lib", "cdylib"]
+crate-type = ["cdylib"]
