+++ programs/sbf/rust/invoked/Cargo.toml
@@ -8,12 +8,9 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[features]
-default = ["program"]
-program = []
-
 [dependencies]
 solana-program = { workspace = true }
+solana-sbf-rust-invoked-dep = { workspace = true }
 
 [lib]
-crate-type = ["lib", "cdylib"]
+crate-type = ["cdylib"]
