+++ programs/sbf/rust/invoked_dep/src/lib.rs
