+++ programs/sbf/rust/realloc_dep/src/lib.rs
