+++ programs/sbf/rust/realloc_invoke/Cargo.toml
@@ -8,13 +8,10 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[features]
-default = ["program"]
-program = []
-
 [dependencies]
 solana-program = { workspace = true }
-solana-sbf-rust-realloc = { workspace = true }
+solana-sbf-rust-realloc-dep = { workspace = true }
+solana-sbf-rust-realloc-invoke-dep = { workspace = true }
 
 [lib]
-crate-type = ["lib", "cdylib"]
+crate-type = ["cdylib"]
