+++ programs/sbf/rust/realloc_invoke_dep/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solana-sbf-rust-sibling-inner-instructions"
+name = "solana-sbf-rust-realloc-invoke-dep"
 version = { workspace = true }
 description = { workspace = true }
 authors = { workspace = true }
@@ -8,8 +8,5 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[dependencies]
-solana-program = { workspace = true }
-
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["lib"]
