+++ programs/sbf/rust/realloc_invoke_dep/src/lib.rs
